Remove Default Date header
It was not possible to remove this header like the Server header. This is an issue when you
need to have full control of the headers that you sent. By the way, it's easy to add this user
with the current headers API. So I don't think this is a good feature. In another hand, we may add an option
like NoDefaultServerHeader but we do not have the *Server where we were writing this header.